ai-service/requirements.txt
# FastAPI AI/ML Service for Event Management Platform
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database & ORM
databases[postgresql]==0.8.0
asyncpg==0.29.0
sqlalchemy==2.0.23

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Machine Learning & Data Science
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Recommendation Engine
surprise==1.1.3
implicit==0.7.2

# API & HTTP
httpx==0.25.2
aiofiles==23.2.1

# Data Validation & Serialization
marshmallow==3.20.1
python-multipart==0.0.6

# Logging & Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Environment & Configuration
python-dotenv==1.0.0
dynaconf==3.2.4

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0

# Code Quality
black==23.10.1
flake8==6.1.0
isort==5.12.0
mypy==1.7.0

# Data Processing
joblib==1.3.2
openpyxl==3.1.2
python-dateutil==2.8.2

# Visualization (for model analysis)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Model Persistence
pickle5==0.0.12
cloudpickle==3.0.0

# Performance
cachetools==5.3.2
memory-profiler==0.61.0
